add 7 homework rework code , add module and arrow function
